new func in cart.js

+++ store/models.py
@@ -12,7 +12,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
 	name = models.CharField(max_length=150, null=True)
-	price = models.FloatField()
+	price = models.DecimalField(max_digits=5, decimal_places=2)
 	digital = models.BooleanField(default=False,null=True, blank=True)
 	images = models.ImageField(null=True, blank=True)
